requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
shap>=0.41.0
pytesseract>=0.3.10
Pillow>=9.0.0
streamlit>=1.11.0
python-dotenv>=0.20.0
joblib>=1.1.0
spacy>=3.3.0
google-generativeai>=0.3.0
python-multipart>=0.0.5
opencv-python-headless>=4.5.0
python-dateutil>=2.8.2
